+++ protected/humhub/modules/admin/messages/cs/views_setting_statistic.php
@@ -1,5 +1,5 @@
 <?php
 return array (
   'Save' => 'Uložit',
-  'You can add an statistics HTML code snippet - which will added to all rendered pags.' => 'Můžete přidat kousek HTML kódu, který se přidá ke všem dosavadním stránkám.',
+  'You can add an statistics HTML code snippet - which will added to all rendered pages.' => 'Můžete přidat kousek HTML kódu, který se přidá ke všem dosavadním stránkám.',
 );