+++ protected/humhub/modules/admin/messages/nl/views_setting_statistic.php
@@ -1,5 +1,5 @@
 <?php
 return array (
   'Save' => 'Bewaar',
-  'You can add an statistics HTML code snippet - which will added to all rendered pags.' => 'U kunt een statistieken HTML code toevoegen die aan alle pagina\'s toegevoegd zal worden.',
+  'You can add an statistics HTML code snippet - which will added to all rendered pages.' => 'U kunt een statistieken HTML code toevoegen die aan alle pagina\'s toegevoegd zal worden.',
 );