+++ protected/humhub/modules/admin/views/setting/statistic.php
@@ -5,7 +5,7 @@ use humhub\compat\CHtml;
 ?>
 <?php $this->beginContent('@admin/views/setting/_advancedLayout.php') ?>
 
-<p><?= Yii::t('AdminModule.views_setting_statistic', 'You can add an statistics HTML code snippet - which will added to all rendered pags.')?></p>
+<p><?= Yii::t('AdminModule.views_setting_statistic', 'You can add an statistics HTML code snippet - which will added to all rendered pages.')?></p>
 <br>
 
 <?php $form = CActiveForm::begin(); ?>