+++ protected/humhub/modules/space/messages/it/views_settings.php
@@ -1,6 +1,6 @@
 <?php
 return array (
   '<strong>Security</strong> settings' => 'Impostazioni <strong>Sicurezza</strong>',
-  '<strong>Space</strong> settings' => '',
+  '<strong>Space</strong> settings' => 'Impostazioni <strong>Spazio</strong>',
   'Permissions are assigned to different user-roles. To edit a permission, select the user-role you want to edit and change the drop-down value of the given permission.' => '',
 );