+++ protected/humhub/modules/user/components/ActiveQueryUser.php
@@ -2,16 +2,18 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2016 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
 namespace humhub\modules\user\components;
 
 use yii\db\ActiveQuery;
+use humhub\modules\user\models\User as UserModel;
+use humhub\events\ActiveQueryEvent;
 
 /**
- * Description of ActiveQueryUser
+ * ActiveQueryUser is used to query User records.
  *
  * @author luke
  */
@@ -19,25 +21,51 @@ class ActiveQueryUser extends ActiveQuery
 {
 
     /**
+     * @event Event an event that is triggered when only visible users are requested via [[visible()]].
+     */
+    const EVENT_CHECK_VISIBILITY = 'checkVisibility';
+
+    /**
+     * @event Event an event that is triggered when only active users are requested via [[active()]].
+     */
+    const EVENT_CHECK_ACTIVE = 'checkActive';
+
+    /**
      * Limit to active users
      * 
-     * @return \humhub\modules\user\components\ActiveQueryUser
+     * @return ActiveQueryUser the query
      */
     public function active()
     {
-        $this->andWhere(['user.status' => \humhub\modules\user\models\User::STATUS_ENABLED]);
+        $this->trigger(self::EVENT_CHECK_ACTIVE, new ActiveQueryEvent(['query' => $this]));
+
+        $this->andWhere(['user.status' => UserModel::STATUS_ENABLED]);
         return $this;
+    }
+
+    /**
+     * Returns only users that should appear in user lists or in the search results.
+     * Also only active (enabled) users are returned.
+     * 
+     * @since 1.2.3
+     * @return ActiveQueryUser the query
+     */
+    public function visible()
+    {
+        $this->trigger(self::EVENT_CHECK_VISIBILITY, new ActiveQueryEvent(['query' => $this]));
+        return $this->active();
     }
 
     /**
      * Adds default user order (e.g. by lastname)
      * 
-     * @return \humhub\modules\user\components\ActiveQueryUser
+     * @return ActiveQueryUser the query
      */
     public function defaultOrder()
     {
         $this->joinWith('profile');
         $this->addOrderBy(['profile.lastname' => SORT_ASC]);
+
         return $this;
     }