+++ protected/humhub/docs/CHANGELOG.md
@@ -40,6 +40,13 @@ Important note for LDAP users: There is a new setting "ID Attribute" which shoul
 - Enh: Added `MobileTargetProvider` abstraction for mobile push notifications
 - Enh: Added `humhub:notification:updateCount` js event
 - Enh: Show space administrators and moderators in member snippet
+- Fix: `humhub\modules\live\Module::getLegitimateContentContainerIds` behaviour with friendship module enabled
+- Enh: Added `BaseNotification:priority` to mark high priority notifications
+- Enh: Added new `User::isVisible` and `ActiveQueryUser::visible` methods
+- Fix: MarkdownEditor cursor position after inserting file/
+- Fix: Make sure own profile content is always visible to user
+- Fix #2501: Do not try to embed Youtube unauthorized videos (acs-ferreira)
+- Fix #2613: Wrong username encoding with pretty url (githubjeka)
 
 1.2.2  (August 2, 2017)
 --------------------------------